% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat_wrapper.R
\name{print.combat_result}
\alias{print.combat_result}
\title{Print Method for \code{combat_result}}
\usage{
\method{print}{combat_result}(x, ...)
}
\arguments{
\item{x}{Object of class \code{combat_result}.}

\item{...}{Unused.}
}
\description{
Print Method for \code{combat_result}
}
