% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_cfg.R
\name{read_cfg}
\alias{read_cfg}
\title{Read a Pipeline Config (JSON)}
\usage{
read_cfg(path)
}
\arguments{
\item{path}{Character. Path to a JSON file.}
}
\value{
A nested list.
}
\description{
Reads a JSON config file used by \code{\link{run_pipeline}}.
}
